chore: Further separate dependencies between test runner and apps.
