+++ tests/Agent/IntegrationTests/Applications/ConsoleInstrumentationLoaderCore/ConsoleInstrumentationLoaderCore.csproj
@@ -20,7 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Shared\Shared.csproj" />
+    <ProjectReference Include="..\..\SharedApplications\Common\SharedApplicationHelpers\SharedApplicationHelpers.csproj" />
   </ItemGroup>
 
 </Project>
