+++ tests/Agent/IntegrationTests/IntegrationTests/IntegrationTests.csproj
@@ -65,9 +65,6 @@
     <ProjectReference Include="..\Models\Models.csproj" />
     <ProjectReference Include="..\Shared\Shared.csproj" />
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
-    <ProjectReference Include="..\SharedApplications\WcfAppIisHosted\WcfAppIisHosted.csproj" />
-  </ItemGroup>
   <Import Project="Test.targets" />
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
