+++ tests/Agent/IntegrationTests/IntegrationTests/ReJit/NetCore/RejitChangeAttributeValue.cs
@@ -46,7 +46,7 @@ namespace NewRelic.Agent.IntegrationTests.ReJit.NetCore
                     _fixture.InitializeApp();
 
                     _fixture.TestChangeAttributeValue();
-                    CommonUtils.ModifyOrCreateXmlAttribute(changeAttributeFilePath, "urn:newrelic-extension", new[] { "extension", "instrumentation", "tracerFactory" }, "metricName", "MyCustomRenamedMetricName");
+                    XmlUtils.ModifyOrCreateXmlAttribute(changeAttributeFilePath, "urn:newrelic-extension", new[] { "extension", "instrumentation", "tracerFactory" }, "metricName", "MyCustomRenamedMetricName");
                     _fixture.AgentLog.WaitForLogLine(AgentLogBase.InstrumentationRefreshFileWatcherComplete, TimeSpan.FromMinutes(1));
                     _fixture.TestChangeAttributeValue();
                 });
