+++ tests/Agent/IntegrationTests/SharedApplications/Common/NetStandardTestLibrary/NetStandardTestLibrary.csproj
@@ -12,7 +12,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Shared\Shared.csproj" />
+    <ProjectReference Include="..\..\..\..\..\..\src\Agent\NewRelic.Api.Agent\NewRelic.Api.Agent.csproj" />
+    <ProjectReference Include="..\SharedApplicationHelpers\SharedApplicationHelpers.csproj" />
   </ItemGroup>
 
 </Project>
