+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/ReflectionHelpers/AssemblyHelper.cs
