+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/Utils/MessageQueueUtil.cs
