+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/WCF/GetCompletedEventArgs.cs
