+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/WCF/IWcfClient.cs
