+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/WCF/IWcfService.cs
