+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/WCF/WcfClient.cs
