+++ tests/Agent/IntegrationTests/SharedApplications/Common/SharedApplicationHelpers/WCF/WcfService.cs
