+++ tests/Agent/IntegrationTests/SharedApplications/WcfAppIisHosted/WcfAppIisHosted.csproj
@@ -49,7 +49,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\IntegrationTestHelpers\IntegrationTestHelpers.csproj" />
+    <ProjectReference Include="..\Common\SharedApplicationHelpers\SharedApplicationHelpers.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
