+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/RemoteServiceFixtures/CouchbaseBasicMvcFixture.cs
@@ -4,11 +4,11 @@
 
 using System;
 using System.Collections.Generic;
-using System.Net;
 using Couchbase;
+using Couchbase.Configuration.Client;
+using Couchbase.Core;
 using NewRelic.Agent.IntegrationTestHelpers.RemoteServiceFixtures;
 using NewRelic.Agent.IntegrationTests.Shared.Couchbase;
-using Xunit;
 
 namespace NewRelic.Agent.UnboundedIntegrationTests.RemoteServiceFixtures
 {
@@ -671,6 +671,45 @@ namespace NewRelic.Agent.UnboundedIntegrationTests.RemoteServiceFixtures
             base.Dispose();
             RemoveDocuments();
             _connection.Dispose();
+        }
+
+        private class CouchbaseConnection : IDisposable
+        {
+            private ICluster _cluster;
+
+            public IBucket Bucket { get; private set; }
+
+            public void Connect()
+            {
+                var config = GetConnectionConfig();
+                _cluster = new Cluster(config);
+                Bucket = _cluster.OpenBucket(CouchbaseTestObject.CouchbaseTestBucket);
+            }
+
+            public void Disconnect()
+            {
+                _cluster.CloseBucket(Bucket);
+                Bucket.Dispose();
+                Bucket = null;
+                _cluster.Dispose();
+                _cluster = null;
+            }
+
+            private ClientConfiguration GetConnectionConfig()
+            {
+                var config = new ClientConfiguration();
+                config.Servers = new List<Uri>()
+            {
+                new Uri(CouchbaseTestObject.CouchbaseServerUrl)
+            };
+                config.UseSsl = false;
+                return config;
+            }
+
+            public void Dispose()
+            {
+                Disconnect();
+            }
         }
     }
 }
